fixed login and altered profile route
